Changed popup to sidebar
