+++ manifest.json
@@ -10,22 +10,11 @@
     },
 
     "permissions": ["tabs", "http://localhost/*"],
-    "content_scripts": [
-        {
-            "matches": ["*://*.youtube.com/*"],
-            "js": ["popup.js"]
-        }
-    ],
 
-    "background": {
-        "scripts": ["src/background.js"]
-    },
-
-    "browser_action": {
+    "sidebar_action": {
         "default_icon": "icons/icon.png",
         "default_title": "Music Downloader",
-        "default_popup": "public/popup.html"
-    },
+        "default_panel": "public/sidebar.html"
+    }
 
-    "content_security_policy": "script-src 'self'; object-src 'self'"
 }
