+++ public/sidebar.html
@@ -3,7 +3,6 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <script src="../src/popup.js"></script>
     <link href="style.css" rel="stylesheet">
     <title>Document</title>
 </head>
@@ -15,7 +14,8 @@
         <label id="title-label" for="title">Title<p class="asterisk">*</p></label>
         <input id="title" name="title" type="text" placeholder="Title" required />
     
-        <button>Submit</button>
+        <button id="submit">Submit</button>
     </form>
+    <script src="../src/sidebar.js"></script>
 </body>
 </html>
